+++ packages/perp/src/types/params.ts
@@ -426,3 +426,16 @@ export interface SimulateScaledLimitOrderResult {
     totalMinSize: BigNumber;
     size: ByBase & ByQuote;
 }
+
+export interface SimulateImpermenantLossParams {
+    expiry: number;
+    instrument: Instrument | InstrumentIdentifier;
+    isInverse: boolean;
+    alphaWadLower: BigNumber;
+    alphaWadUpper: BigNumber;
+}
+
+export interface SimulateImpermenantLossResult {
+    tick: number;
+    impermanentLoss: number;
+}